feat : get post data from server& display
