+++ src/pages/Survey/Post.css
